+++ luto/solvers/solver.py
@@ -1022,8 +1022,8 @@ class LutoSolver:
             self.gurobi_model.remove(
                 list(self.X_ag_irr_vars_jr[:, r][np.where(self.X_ag_irr_vars_jr[:, r])])
             )
-            self.X_ag_dry_vars_jr[:, r] = np.zeros(self._input_data.n_ag_lus)
-            self.X_ag_irr_vars_jr[:, r] = np.zeros(self._input_data.n_ag_lus)
+            self.X_ag_dry_vars_jr[:, r] = np.zeros(self._input_data.n_ag_lus).astype(np.float32)
+            self.X_ag_irr_vars_jr[:, r] = np.zeros(self._input_data.n_ag_lus).astype(np.float32)
             for j in range(self._input_data.n_ag_lus):
                 if self._input_data.ag_x_mrj[0, r, j]:
                     self.X_ag_dry_vars_jr[j, r] = self.gurobi_model.addVar(
@@ -1039,7 +1039,7 @@ class LutoSolver:
             self.gurobi_model.remove(
                 list(self.X_non_ag_vars_kr[:, r][np.where(self.X_non_ag_vars_kr[:, r])])
             )
-            self.X_non_ag_vars_kr[:, r] = np.zeros(self._input_data.n_non_ag_lus)
+            self.X_non_ag_vars_kr[:, r] = np.zeros(self._input_data.n_non_ag_lus).astype(np.float32)
             for k, non_ag_lu_desc in zip(
                 range(self._input_data.n_non_ag_lus), NON_AG_LAND_USES
             ):
@@ -1075,8 +1075,8 @@ class LutoSolver:
                         ]
                     )
                 )
-                self.X_ag_man_dry_vars_jr[am][:, r] = np.zeros(len(am_j_list))
-                self.X_ag_man_irr_vars_jr[am][:, r] = np.zeros(len(am_j_list))
+                self.X_ag_man_dry_vars_jr[am][:, r] = np.zeros(len(am_j_list)).astype(np.float32)
+                self.X_ag_man_irr_vars_jr[am][:, r] = np.zeros(len(am_j_list)).astype(np.float32)
 
             for m, j in self._input_data.cells2ag_lu[r]:
                 # replace am variables
